+++ poem-openapi/src/payload/binary.rs
@@ -53,7 +53,7 @@ impl<T: Into<Vec<u8>> + Send> ApiResponse for Binary<T> {
     fn meta() -> MetaResponses {
         MetaResponses {
             responses: vec![MetaResponse {
-                description: None,
+                description: "",
                 status: Some(200),
                 content: vec![MetaMediaType {
                     content_type: Self::CONTENT_TYPE,