+++ poem-openapi/src/payload/json.rs
@@ -57,7 +57,7 @@ impl<T: ToJSON> ApiResponse for Json<T> {
     fn meta() -> MetaResponses {
         MetaResponses {
             responses: vec![MetaResponse {
-                description: None,
+                description: "",
                 status: Some(200),
                 content: vec![MetaMediaType {
                     content_type: Self::CONTENT_TYPE,